Rewrite based on the code of the nmap integration
Devices are now tracked based on the MAC, not IP.  The tracker reports
the IP, the MAC and the name (if available) from the ARP tables.  We use
arp -a to list the devices currently visible, ping them once each, and
report to HA if ping reply received.  Seems to work reliably with my
Android phone.  The config options are now more like the nmap options
(hosts and exclude) but I kept the interface option from the arp+ping
component.  We ping all the devices found in the arp tables rather than
needing a list of devices to track, except for devices outside of the
"hosts" list or matching the "exclude" list.  The format now allows
netmasks like 10.0.0.0/16, similar to nmap.