Try an ARP query if ping fails
Some Windows laptops, probably those running firewalls, don't reply to
ICMP pings so try an ARP ping too.  Use the nping utility for this
because it doesn't ask for the interface.  arping wants us to pass -I
with the interface name in addition to the target ip and we'd have to
look through the routing table.